requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0

# Document parsing
python-docx==1.1.0

# Browser automation
playwright==1.40.0

# Task queue (optional for async processing)
celery==5.3.4
redis==5.0.1

# Utilities
Werkzeug==3.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
